+++ commonjs/documentation/changelog.md
@@ -1,5 +1,8 @@
 # Change Log
 <pre>
+v3.2.1  Added GeoFences APIs(create, update, query, remove) [TIMODOPEN-349]
+        Added PushNotifications APIs(setBadge, resetBadge) [TIMODOPEN-354]
+
 v3.2.0  Bumping version to match the TiSDK version this module is being shipped with
 
 v2.3.8  Updated the example app to use the Facebook module since it was pulled out of the SDK.[TIMODOPEN-276]